+++ tests/preprocessing/test_preprocessor_normalizer.py
@@ -40,6 +40,37 @@ def input_normalizer():
     return InputNormalizer(config=config.data.normalizer, data_indices=data_indices, statistics=statistics)
 
 
+@pytest.fixture()
+def remap_normalizer():
+    config = DictConfig(
+        {
+            "diagnostics": {"log": {"code": {"level": "DEBUG"}}},
+            "data": {
+                "normalizer": {
+                    "default": "mean-std",
+                    "remap": {"x": "z", "y": "x"},
+                    "min-max": ["x"],
+                    "max": ["y"],
+                    "none": ["z"],
+                    "mean-std": ["q"],
+                },
+                "forcing": ["z", "q"],
+                "diagnostic": ["other"],
+                "remapped": {},
+            },
+        },
+    )
+    statistics = {
+        "mean": np.array([1.0, 2.0, 3.0, 4.5, 3.0]),
+        "stdev": np.array([0.5, 0.5, 0.5, 1, 14]),
+        "minimum": np.array([1.0, 1.0, 1.0, 1.0, 1.0]),
+        "maximum": np.array([11.0, 10.0, 10.0, 10.0, 10.0]),
+    }
+    name_to_index = {"x": 0, "y": 1, "z": 2, "q": 3, "other": 4}
+    data_indices = IndexCollection(config=config, name_to_index=name_to_index)
+    return InputNormalizer(config=config.data.normalizer, statistics=statistics, data_indices=data_indices)
+
+
 def test_normalizer_not_inplace(input_normalizer) -> None:
     x = torch.Tensor([[1.0, 2.0, 3.0, 4.0, 5.0], [6.0, 7.0, 8.0, 9.0, 10.0]])
     input_normalizer(x, in_place=False)
@@ -87,3 +118,15 @@ def test_normalize_inverse_transform(input_normalizer) -> None:
     assert torch.allclose(
         input_normalizer.inverse_transform(input_normalizer.transform(x, in_place=False), in_place=False), x
     )
+
+
+def test_normalizer_not_inplace_remap(remap_normalizer) -> None:
+    x = torch.Tensor([[1.0, 2.0, 3.0, 4.0, 5.0], [6.0, 7.0, 8.0, 9.0, 10.0]])
+    remap_normalizer(x, in_place=False)
+    assert torch.allclose(x, torch.Tensor([[1.0, 2.0, 3.0, 4.0, 5.0], [6.0, 7.0, 8.0, 9.0, 10.0]]))
+
+
+def test_normalize_remap(remap_normalizer) -> None:
+    x = torch.Tensor([[1.0, 2.0, 3.0, 4.0, 5.0], [6.0, 7.0, 8.0, 9.0, 10.0]])
+    expected_output = torch.Tensor([[0.0, 2 / 11, 3.0, -0.5, 1 / 7], [5 / 9, 7 / 11, 8.0, 4.5, 0.5]])
+    assert torch.allclose(remap_normalizer.transform(x), expected_output)